+++ AddOns/XIV_Databar_Continued/Core/changelog/3.1.lua
@@ -1,0 +1,48 @@
+local AddOnName, XIVBar = ...;
+
+XIVBar.Changelog[3100] = {
+    version_string = "3.1",
+    release_date = "2024/09/02",
+    important = {
+        ["zhCN"] = {},
+        ["zhTW"] = {},
+        ["enUS"] = {
+            "[Retail] [Travel] module fix (Acherus, Dreamgrove, ...) for Retail's API changes in 11.0.2.",
+            "[Cataclysm Classic] Added Archaeology in the [Tradeskill] module (Tradeskill tooltip).",
+            "[Cataclysm Classic] Added back the Right Click functionality in the [Tradeskill] module to make it possible to access the professions frame.",
+            "[Classic SoD] [Tradeskill] module fix.",
+        },
+        ["frFR"] = {
+            "[Retail] Correctif (Acherus, Sylverêve, ...) du module [Voyage] pour prendre en compte les changements d'API de la version 11.0.2 de Retail.",
+            "[Cataclysm Classic] Ajout de l'Archéologie dans le module [Tradeskill] (infobulle des professions).",
+            "[Cataclysm Classic] Ajout de la fonctionnalité de clique droit dans le module [Tradeskill] pour pouvoir accéder au module de la fiche de la profession.",
+            "[Classic SoD] Correction du module [Professions].",
+        },
+        ["koKR"] = {},
+        ["ruRU"] = {}
+    },
+    new = {
+        ["zhCN"] = {},
+        ["zhTW"] = {},
+        ["enUS"] = {
+            "[Retail] New [Mythic+ Portals] functionality in the [Travel] module, allowing you to have fast access to your known [Mythic+ Portals] and to cast that teleport.",
+        },
+        ["frFR"] = {
+            "[Retail] Nouvelle fonctionnalité [Portails Mythique+] dans le module [Voyage], permettant d'avoir rapidement accès aux [Portails Mythique+] connus et de lancer la téléportation directement.",
+        },
+        ["koKR"] = {},
+        ["ruRU"] = {}
+    },
+    improvment = {
+        ["zhCN"] = {},
+        ["zhTW"] = {},
+        ["enUS"] = {
+            "[Global] Code refactoring to fix underscores reassignments causing erros with some native Blizzard commands in XIV Databr Continued.",
+        },
+        ["frFR"] = {
+            "[Global] Réadaptation du code pour corriger les réassignations d'underscores causant des erreurs avec certaines commandes natives de Blizzard dans XIV Databar Continued.",
+        },
+        ["koKR"] = {},
+        ["ruRU"] = {}
+    }
+}