+++ AddOns/XIV_Databar_Continued/locales/zhCN.lua
@@ -151,10 +151,25 @@ L['Available Ranks'] = "神器等级";
 L['Artifact Knowledge'] = "神器知识";
 
 -- Travel (Translation needed)
+L["Mythic+ Portals"] = true;
+L['Show Mythic+ Portals'] = true;
 L['Use Random Hearthstone'] = true;
 L['Empty Hearthstones List'] = "If you see an empty list, /reload your UI a few seconds after the initial loading (Blizzard is loading items informations asynchronously so that's the only solution for now)."
 L['Hearthstones Select'] = true;
 L['Hearthstones Select Desc'] = "Select which hearthstones to use (be careful if you select multiple hearthstones, you might want to check the 'Hearthstones Select' option)";
+
+L["Classic"] = true;
+L["Burning Crusade"] = true;
+L["Wrath of the Lich King"] = true;
+L["Cataclysm"] = true;
+L["Mists of Pandaria"] = true;
+L["Warlords of Draenor"] = true;
+L["Legion"] = true;
+L["Battle for Azeroth"] = true;
+L["Shadowlands"] = true;
+L["Dragonflight"] = true;
+L["The War Within"] = true;
+L["Current season"] = true;
 
 -- Changelog
 L["%month%-%day%-%year%"] = "%year%年%month%月%day%日";