+++ AddOns/XIV_Databar_Continued/CHANGELOG.md
@@ -1,10 +1,23 @@
 # XIV_Databar Continued
 
-## [3.0.3](https://github.com/ZelionGG/XIV_Databar-Continued/tree/v3.0.3) (2024-08-26)
+## [3.1](https://github.com/ZelionGG/XIV_Databar-Continued/tree/v3.1) (2024-09-03)
 
-[Full Changelog](https://github.com/ZelionGG/XIV_Databar-Continued/compare/v3.0.2...v3.0.3) [Previous Releases](https://github.com/ZelionGG/XIV_Databar-Continued/releases)
+[Full Changelog](https://github.com/ZelionGG/XIV_Databar-Continued/compare/v3.0.3...v3.1) [Previous Releases](https://github.com/ZelionGG/XIV_Databar-Continued/releases)
 
 ### _Retail :_
 
-- Micromenu module chat (bubble) button fix for Retail 11.0.2.
-- Travel module fix for Retail's API changes in 11.0.2.
+- New Mythic+ Portals functionality in the Travel module, allowing you to have fast access to your known Mythic+ Portals and to cast that teleport.
+- Travel module fix (Acherus, Dreamgrove, ...) for Retail's API changes in 11.0.2.
+
+### _Cataclysm Classic :_
+
+- Added Archaeology in the Tradeskill module tooltip.
+- Added back the Right Click functionality in the Tradeskill module to make it possible to access the professions frame.
+
+### _Classic SoD :_
+
+- Tradeskill module fix.
+
+### _Global :_
+
+- Code refactoring to fix underscores reassignments causing erros with some native Blizzard commands in XIV Databr Continued.